+++ src/bin/mdcat/args.rs
@@ -21,19 +21,23 @@ Report issues to <https://codeberg.org/flausch/mdcat>.",
             Arg::new("paginate")
                 .short('p')
                 .long("--paginate")
+                .action(ArgAction::SetTrue)
                 .help("Paginate the output of mdcat with a pager like less."),
         )
         .arg(
             Arg::new("no_pager")
                 .short('P')
                 .long("--no-pager")
+                .action(ArgAction::SetTrue)
                 .help("Do not page output.  Default if invoked as mdcat")
                 .overrides_with("paginate"),
         )
         .arg(
             Arg::new("filenames")
-                .multiple_occurrences(true)
+                .action(ArgAction::Append)
+                .multiple_values(true)
                 .help("The file to read.  If - read from standard input instead")
+                .value_parser(clap::builder::NonEmptyStringValueParser::new())
                 .value_hint(ValueHint::FilePath)
                 .default_value("-"),
         )
@@ -42,11 +46,13 @@ Report issues to <https://codeberg.org/flausch/mdcat>.",
                 .short('c')
                 .long("--no-colour")
                 .aliases(&["nocolour", "no-color", "nocolor"])
+                .action(ArgAction::SetTrue)
                 .help("Disable all colours and other styles."),
         )
         .arg(
             Arg::new("columns")
                 .long("columns")
+                .value_parser(clap::value_parser!(usize))
                 .help("Maximum number of columns to use for output")
                 .default_value(default_columns),
         )
@@ -54,16 +60,19 @@ Report issues to <https://codeberg.org/flausch/mdcat>.",
             Arg::new("local_only")
                 .short('l')
                 .long("local")
+                .action(ArgAction::SetTrue)
                 .help("Do not load remote resources like images"),
         )
         .arg(
             Arg::new("fail_fast")
                 .long("fail")
+                .action(ArgAction::SetTrue)
                 .help("Exit immediately if any error occurs processing an input file"),
         )
         .arg(
             Arg::new("detect_only")
                 .long("detect-only")
+                .action(ArgAction::SetTrue)
                 .help("Only detect the terminal type and exit")
                 .hide(true),
         )
@@ -71,6 +80,7 @@ Report issues to <https://codeberg.org/flausch/mdcat>.",
             Arg::new("ansi_only")
                 .long("ansi-only")
                 .help("Limit to standard ANSI formatting")
+                .action(ArgAction::SetTrue)
                 .conflicts_with("no_colour")
                 .hide(true),
         )